Complete file restructure, fixing issues
