+++ client/src/components/assets/pictures/explore/.picasa.ini
