+++ client/src/components/layout/Home.js
@@ -1,13 +1,14 @@
 import React from 'react';
-import Carousel from '../presentational/Carousel/Carousel';
-import Sitemap from '../presentational/sitemap/Sitemap';
 import Hero from '../presentational/Hero/Hero';
-
+import Carousel from '../presentational/Carousel/Carousel';
+import Cards from '../presentational/Cards/Cards';
+import Sitemap from '../presentational/Sitemap/Sitemap';
 
 const Home = () => (
   <div>
     <Hero />
     <Carousel />
+    <Cards />
     <Sitemap />
   </div>
 );
