--- src/App.css
@@ -1,74 +1,0 @@
-.sitemap_container {
-  border-top: 1px solid #D1D1D1;  
-  margin-bottom: 10px;
-  
-}
-.sitemap {
-  display: flex;
-  padding: 50px 20px 40px 20px;
-  width: 100%;
-  justify-content: space-between;
-  border-bottom: 1px solid #D1D1D1;
-}
-
-@media (min-width: 1130px) {
-  .sitemap {
-    width: 900px;
-    margin: auto;
-  }
-}
-
-.sitemap article {
-  margin-top: 0px;
-  margin-right: 10px;
-  padding-top: 0px;
-  text-align: left;  
-}
-
-.sitemap_lists {
-  padding-left: 60px;
-  flex-basis: 24%;
-}
-
-.sitemap article h4 {
-  margin-top: 0px;
-  font-size: 0.9rem;  
-  color: #4D4A4A;  
-}
-
-.sitemap article ul {
-  list-style-type: none;
-  text-align: left;  
-  padding-left: 0px;
-  font-size: 0.9rem;
-  line-height: 1.25rem;
-  color: #5d5a5a;
-}
-
-.sitemap_regional {
-  flex-grow: 2;
-  display: flex;
-  flex-direction: column;
-}
-
-.sitemap_regional select {
-  font-size: 1.1rem;
-  color: #5d5a5a;
-  padding: 12px;
-  margin-bottom: 16px;
-  width: 100%;
-}
-
-@media (max-width: 750px) {
-  .sitemap_lists {
-    display: none;
-  }
-
-  .sitemap_regional {
-    flex-direction: row;
-  }
-
-  .sitemap_regional select:first-child {
-    margin-right: 20px;
-  }
-}
