--- src/components/Sitemap/styles.js
@@ -1,5 +1,0 @@
-import styled from 'styled-components';
-import theme from '../../stylesConfig';
-
-
-
